refactor: move accordion item services closer to use sites
