+++ containers/ecr-viewer/src/app/view-data/types.ts
@@ -1,0 +1,3 @@
+import { Accordion } from "@trussworks/react-uswds";
+
+export type AccordionItem = React.ComponentProps<typeof Accordion>["items"][0];
